# Optimizations in escape hatch