--- a/scalafix-core/src/main/scala/scalafix/internal/patch/EscapeHatch.scala
+++ b/scalafix-core/src/main/scala/scalafix/internal/patch/EscapeHatch.scala
@@ -9,6 +9,7 @@ import scala.meta.contrib._
 import scala.meta.tokens.Token
 import scala.meta.tokens.Token.Comment
 import scalafix.internal.config.FilterMatcher
+import scalafix.internal.diff.DiffDisable
 import scalafix.internal.patch.EscapeHatch._
 import scalafix.internal.util.LintSyntax._
 import scalafix.lint.RuleDiagnostic
@@ -23,13 +24,13 @@ import scalafix.v0._
  * precedence over the former in case there are overlaps.
  * See `AnchoredEscapes` and `AnnotatedEscapes` for more details.
  */
-class EscapeHatch private (
+private[scalafix] class EscapeHatch private (
     anchoredEscapes: AnchoredEscapes,
-    annotatedEscapes: AnnotatedEscapes) {
+    annotatedEscapes: AnnotatedEscapes,
+    diffDisable: DiffDisable) {
 
-  def rawFilter(
-      patchesByName: Map[RuleName, Patch],
-      ctx: RuleCtx): (Patch, List[RuleDiagnostic]) = {
+  private def rawFilter(patchesByName: Map[RuleName, Patch])(
+      implicit ctx: RuleCtx): (Patch, List[RuleDiagnostic]) = {
     var patchBuilder = Patch.empty
     val diagnostics = List.newBuilder[RuleDiagnostic]
     patchesByName.foreach {
@@ -44,16 +45,11 @@ class EscapeHatch private (
     (patchBuilder, diagnostics.result())
   }
 
-  def filter(
-      patchesByName: Map[RuleName, Patch],
-      ctx: RuleCtx,
-      index: SemanticdbIndex
-  ): (Patch, List[RuleDiagnostic]) = {
-    if (!FastPatch.hasSuppression(ctx.input.text) && ctx.diffDisable.isEmpty) {
-      return rawFilter(patchesByName, ctx)
-    }
-    val diff = ctx.diffDisable
-    val config = ctx.config
+  def filter(patchesByName: Map[RuleName, Patch])(
+      implicit ctx: RuleCtx,
+      index: SemanticdbIndex): (Patch, List[RuleDiagnostic]) = {
+    if (isEmpty) return rawFilter(patchesByName)
+
     val usedEscapes = mutable.Set.empty[EscapeFilter]
     val lintMessages = List.newBuilder[RuleDiagnostic]
 
@@ -71,30 +67,24 @@ class EscapeHatch private (
 
     def loop(name: RuleName, patch: Patch): Patch = patch match {
       case AtomicPatch(underlying) =>
-        val hasDisabledPatch = {
-          val patches = PatchInternals.treePatchApply(underlying)(ctx, index)
-          patches.exists { tp =>
-            val byGit = diff.isDisabled(tp.tok.pos)
+        val hasDisabledPatch =
+          PatchInternals.treePatchApply(underlying).exists { tp =>
+            val byGit = diffDisable.isDisabled(tp.tok.pos)
             val byEscape = isDisabledByEscape(name, tp.tok.pos.start)
             byGit || byEscape
           }
-        }
+        if (hasDisabledPatch) EmptyPatch else underlying
 
-        if (hasDisabledPatch) EmptyPatch
-        else loop(name, underlying)
-
-      case Concat(a, b) =>
-        Concat(loop(name, a), loop(name, b))
+      case Concat(a, b) => Concat(loop(name, a), loop(name, b))
 
       case LintPatch(lint) =>
-        val byGit = diff.isDisabled(lint.position)
+        val byGit = diffDisable.isDisabled(lint.position)
         val id = lint.fullStringID(name)
         val byEscape = isDisabledByEscape(id, lint.position.start)
-
         val isLintDisabled = byGit || byEscape
 
         if (!isLintDisabled) {
-          lintMessages += lint.toDiagnostic(name, config)
+          lintMessages += lint.toDiagnostic(name, ctx.config)
         }
 
         EmptyPatch
@@ -108,20 +98,23 @@ class EscapeHatch private (
     val unusedWarnings =
       (annotatedEscapes.unusedEscapes(usedEscapes) ++
         anchoredEscapes.unusedEscapes(usedEscapes)).map { pos =>
-        UnusedScalafixSuppression.at(pos).toDiagnostic(UnusedName, config)
+        UnusedScalafixSuppression.at(pos).toDiagnostic(UnusedName, ctx.config)
       }
     val warnings = lintMessages.result() ++ unusedWarnings
     (patches, warnings)
   }
+
+  def isEmpty: Boolean =
+    diffDisable.isEmpty && anchoredEscapes.isEmpty && annotatedEscapes.isEmpty
 }
 
 object EscapeHatch {
-
   private val UnusedScalafixSuppression =
     LintCategory.warning("", "Unused Scalafix suppression, this can be removed")
   private val UnusedName = RuleName("UnusedScalafixSuppression")
 
   private type EscapeTree = TreeMap[EscapeOffset, List[EscapeFilter]]
+  private val EmptyEscapeTree: EscapeTree = TreeMap.empty
 
   private case class EscapeFilter(
       matcher: FilterMatcher,
@@ -132,17 +125,23 @@ object EscapeHatch {
       id.identifiers.exists(i => matcher.matches(i.value))
   }
 
-  private case class EscapeOffset(offset: Int)
+  private case class EscapeOffset(offset: Int) extends AnyVal
 
   private object EscapeOffset {
     implicit val ordering: Ordering[EscapeOffset] = Ordering.by(_.offset)
   }
 
-  def apply(tree: Tree, associatedComments: AssociatedComments): EscapeHatch =
+  def apply(
+      input: Input,
+      tree: => Tree,
+      associatedComments: => AssociatedComments,
+      diffDisable: DiffDisable): EscapeHatch = {
     new EscapeHatch(
-      AnchoredEscapes(tree, associatedComments),
-      AnnotatedEscapes(tree)
+      AnchoredEscapes(input, tree, associatedComments),
+      AnnotatedEscapes(input, tree),
+      diffDisable
     )
+  }
 
   /** Rules are disabled via standard `@SuppressWarnings` annotations. The
    * annotation can be placed in class, object, trait, type, def, val, var,
@@ -157,18 +156,23 @@ object EscapeHatch {
   private class AnnotatedEscapes private (escapeTree: EscapeTree) {
     import AnnotatedEscapes._
 
+    val isEmpty: Boolean = escapeTree.isEmpty
+
     def isEnabled(
         ruleName: RuleName,
         position: Int): (Boolean, Option[EscapeFilter]) = {
-      val escapesUpToPos =
-        escapeTree.to(EscapeOffset(position)).valuesIterator.flatten
-      escapesUpToPos
-        .collectFirst {
-          case f @ EscapeFilter(_, _, _, Some(end))
-              if f.matches(ruleName) && end.offset >= position =>
-            (false, Some(f))
-        }
-        .getOrElse(true -> None)
+      if (isEmpty) (true, None)
+      else {
+        val escapesUpToPos =
+          escapeTree.to(EscapeOffset(position)).valuesIterator.flatten
+        escapesUpToPos
+          .collectFirst {
+            case f @ EscapeFilter(_, _, _, Some(end))
+                if end.offset >= position && f.matches(ruleName) =>
+              (false, Some(f))
+          }
+          .getOrElse(true -> None)
+      }
     }
 
     def unusedEscapes(used: collection.Set[EscapeFilter]): Set[Position] =
@@ -184,7 +188,12 @@ object EscapeHatch {
     private val SuppressAll = "all"
     private val OptionalRulePrefix = "scalafix:"
 
-    def apply(tree: Tree): AnnotatedEscapes = {
+    def apply(input: Input, tree: => Tree): AnnotatedEscapes = {
+      if (input.text.contains(SuppressWarnings)) collectEscapes(tree)
+      else new AnnotatedEscapes(EmptyEscapeTree)
+    }
+
+    private def collectEscapes(tree: Tree): AnnotatedEscapes = {
       val escapes =
         tree.collect {
           case t @ Mods(SuppressWarningsArgs(args)) =>
@@ -239,7 +248,8 @@ object EscapeHatch {
       }
   }
 
-  /** Rules are disabled via comments with a specific syntax:
+  /**
+   * Rules are disabled via comments with a specific syntax:
    * `scalafix:off` or `scalafix:on` disable or enable rules until the end of file
    * `scalafix:ok` disable rules on an associated expression
    * a list of rules separated by commas can be provided to selectively
@@ -253,6 +263,8 @@ object EscapeHatch {
       enabling: EscapeTree,
       disabling: EscapeTree,
       unused: List[Position]) {
+
+    val isEmpty: Boolean = disabling.isEmpty && unused.isEmpty
 
     /**
      * a rule r is disabled in position p if there is a comment disabling r at
@@ -269,7 +281,7 @@ object EscapeHatch {
       @tailrec
       def loop(
           disables: List[EscapeFilter],
-          culprit: Option[EscapeFilter]): (Boolean, Option[EscapeFilter]) =
+          culprit: Option[EscapeFilter]): (Boolean, Option[EscapeFilter]) = {
         disables match {
           case head :: tail if head.matches(ruleName) =>
             head.endOffset match {
@@ -285,9 +297,14 @@ object EscapeHatch {
           case _ :: tail => loop(tail, culprit)
           case Nil => (true, culprit)
         }
+      }
 
-      val disables = disabling.to(EscapeOffset(position)).valuesIterator.flatten
-      loop(disables.toList, None)
+      if (disabling.isEmpty) (true, None)
+      else {
+        val disables =
+          disabling.to(EscapeOffset(position)).valuesIterator.flatten
+        loop(disables.toList, None)
+      }
     }
 
     def unusedEscapes(used: collection.Set[EscapeFilter]): Set[Position] =
@@ -296,16 +313,29 @@ object EscapeHatch {
   }
 
   private object AnchoredEscapes {
-    private val FilterDisable = regex("off")
-    private val FilterEnable = regex("on")
-    private val FilterExpression = regex("ok")
+    private val Prefix = "scalafix:"
+    private val ScalafixOk = Prefix + "ok"
+    private val ScalafixOn = Prefix + "on"
+    private val ScalafixOff = Prefix + "off"
+    private val ScalafixOkRgx = regex(ScalafixOk)
+    private val ScalafixOnRgx = regex(ScalafixOn)
+    private val ScalafixOffRgx = regex(ScalafixOff)
 
-    private def regex(anchor: String) =
-      s"\\s?scalafix:$anchor\\s?(.*?)(?:;.*)?".r
+    private def regex(anchor: String) = s"\\s?$anchor\\s?(.*?)(?:;.*)?".r
 
     def apply(
-        tree: Tree,
-        associatedComments: AssociatedComments): AnchoredEscapes = {
+        input: Input,
+        tree: => Tree,
+        associatedComments: => AssociatedComments): AnchoredEscapes = {
+      if (input.text.contains(Prefix))
+        collectEscapes(input, tree, associatedComments)
+      else new AnchoredEscapes(EmptyEscapeTree, EmptyEscapeTree, Nil)
+    }
+
+    private def collectEscapes(
+        input: Input,
+        tree: => Tree,
+        associatedComments: => AssociatedComments): AnchoredEscapes = {
       val enable = TreeMap.newBuilder[EscapeOffset, List[EscapeFilter]]
       val disable = TreeMap.newBuilder[EscapeOffset, List[EscapeFilter]]
       val visitedFilterExpression = mutable.Set.empty[Position]
@@ -349,25 +379,27 @@ object EscapeHatch {
         rulesToPos.result()
       }
 
-      tree.foreach { t =>
-        associatedComments.trailing(t).foreach {
-          // matches simple expressions
-          //
-          // val a = (
-          //   1,
-          //   2
-          // ) // scalafix:ok RuleA
-          //
-          case comment @ Token.Comment(FilterExpression(rules)) =>
-            if (!visitedFilterExpression.contains(comment.pos)) {
-              val rulesPos = rulesWithPosition(rules, comment)
-              val start = EscapeOffset(t.pos.start)
-              val end = Some(EscapeOffset(t.pos.end))
-              disable += (start -> makeFilters(rulesPos, start, end, comment))
-              visitedFilterExpression += comment.pos
-            }
+      if (input.text.contains(ScalafixOk)) {
+        tree.foreach { t =>
+          associatedComments.trailing(t).foreach {
+            // matches simple expressions
+            //
+            // val a = (
+            //   1,
+            //   2
+            // ) // scalafix:ok RuleA
+            //
+            case comment @ Token.Comment(ScalafixOkRgx(rules)) =>
+              if (!visitedFilterExpression.contains(comment.pos)) {
+                val rulesPos = rulesWithPosition(rules, comment)
+                val start = EscapeOffset(t.pos.start)
+                val end = Some(EscapeOffset(t.pos.end))
+                disable += (start -> makeFilters(rulesPos, start, end, comment))
+                visitedFilterExpression += comment.pos
+              }
 
-          case _ => ()
+            case _ => ()
+          }
         }
       }
 
@@ -379,7 +411,7 @@ object EscapeHatch {
             // // scalafix:off RuleA
             // ...
             //
-            case FilterDisable(rules) =>
+            case ScalafixOffRgx(rules) =>
               val rulesPos = rulesWithPosition(rules, comment)
               val start = EscapeOffset(comment.pos.start)
               disable += (start -> makeFilters(rulesPos, start, None, comment))
@@ -390,7 +422,7 @@ object EscapeHatch {
             // ...
             // // scalafix:on RuleA
             //
-            case FilterEnable(rules) =>
+            case ScalafixOnRgx(rules) =>
               val rulesPos = rulesWithPosition(rules, comment)
               val start = EscapeOffset(comment.pos.start)
               enable += (start -> makeFilters(rulesPos, start, None, comment))
@@ -401,7 +433,7 @@ object EscapeHatch {
             // object Dummy { // scalafix:ok NoDummy
             //   1
             // }
-            case FilterExpression(rules) =>
+            case ScalafixOkRgx(rules) =>
               if (!visitedFilterExpression.contains(comment.pos)) {
                 val rulesPos = rulesWithPosition(rules, comment)
                 // we approximate the position of the expression to the whole line