--- a/scalafix-core/src/main/scala/scalafix/internal/v1/InternalDoc.scala
+++ b/scalafix-core/src/main/scala/scalafix/internal/v1/InternalDoc.scala
@@ -4,7 +4,6 @@ import scala.meta.Tree
 import scala.meta.contrib.AssociatedComments
 import scala.meta.inputs.Input
 import scalafix.internal.config.ScalafixConfig
-import scalafix.internal.diff.DiffDisable
 import scalafix.internal.patch.EscapeHatch
 import scalafix.util.MatchingParens
 import scalafix.util.TokenList
@@ -15,7 +14,6 @@ class InternalDoc(
     val comments: LazyValue[AssociatedComments],
     val config: ScalafixConfig,
     val escapeHatch: LazyValue[EscapeHatch],
-    val diffDisable: DiffDisable,
     val matchingParens: LazyValue[MatchingParens],
     val tokenList: LazyValue[TokenList]
 )