--- a/scalafix-core/src/main/scala/scalafix/v0/RuleCtx.scala
+++ b/scalafix-core/src/main/scala/scalafix/v0/RuleCtx.scala
@@ -50,7 +50,6 @@ trait RuleCtx extends PatchOps {
   private[scalafix] def toks(t: Tree): Tokens
   private[scalafix] def config: ScalafixConfig
   private[scalafix] def escapeHatch: EscapeHatch
-  private[scalafix] def diffDisable: DiffDisable
 }
 
 object RuleCtx {
@@ -64,5 +63,5 @@ object RuleCtx {
       tree: Tree,
       config: ScalafixConfig,
       diffDisable: DiffDisable): RuleCtx =
-    RuleCtxImpl(tree, config, diffDisable)
+    new RuleCtxImpl(tree, config, diffDisable)
 }