--- a/scalafix-core/src/main/scala/scalafix/v1/SyntacticDocument.scala
+++ b/scalafix-core/src/main/scala/scalafix/v1/SyntacticDocument.scala
@@ -63,7 +63,7 @@ object SyntacticDocument {
       AssociatedComments(tree.value)
     }
     val escape = LazyValue.later { () =>
-      EscapeHatch(input, tree.value, comments.value, diffDisable)
+      EscapeHatch(input, tree, comments, diffDisable)
     }
     val matchingParens = LazyValue.later { () =>
       MatchingParens(tokens.value)