… scalafix-core/src/main/scala/scalafix/internal/patch/FastPatch.scala