This PR implements some optimizations in `EscapeHatch` (see details in https://github.com/scalacenter/scalafix/issues/894):

- short circuit the parsing logic when there is no suppression in a given file (each suppression method  is short circuited individually). This avoids the cost of traversing the scalameta `Tree`.
- Make atomic empty patches be considered empty. This will potentially maximize the cases where parts or the entire `EscapeHatch` is short circuited
- A few other minor improvements

The tables below compare the running time of what is on `master` vs. this PR + https://github.com/scalameta/scalameta/pull/1793. These figures are the median of 5 consecutive runs of Scalafix against the Spark code base (3201 files; 103181 lines). The setup is the same as explained [here](https://github.com/scalacenter/scalafix/issues/894#issue-363820370):
>The rule utilized for this exercise is purposely simple to reduce noise and make `EscapeHatch` stand out. Different rules will give different results.

- Without suppression:

|        | Scalafix (total)   | `EscapeHatch` (total) | `:ok`         | `:on\|:off`   | `@`          | `filter`      | 
| ------ | ------- | ------------- | ------------- | ------------- | ------------ | ------------- |
| Before  | 50354ms | 17454ms (34%) | 10326ms (20%) | 252ms (0.5%)  | 6795ms (13%) | 0ms (0%) |
| After | 26066ms | 61ms (0%)     | 0ms (0%)      | 0ms (0%)      | 18ms (0%)    | 0ms (0%)  |


- With suppression (all flavors put together in every source file):

|        | Scalafix (total)   | `EscapeHatch` (total) | `:ok`         | `:on\|:off`   | `@`          | `filter`      | 
| ------ | ------- | ------------- | ------------- | ------------- | ------------ | ------------- |
| Before   | 55094ms | 27479ms (50%) | 9885ms (18%) | 342ms (0.6%)  | 6470ms (11%) | 10718ms (19%) |
| After | 30598ms | 11456ms (37%) | 6330ms (20%) | 356ms (1%)    | 2364ms (7%)    | 2362ms (7%)  |


I believe that most projects using Scalafix don't use suppressions often (only exceptionally as it should be). For those, the cost of `EscapeHatch` is now zero. Any source file containing suppressions only pay the price of the chosen suppression method.

As highlighted [here](https://github.com/scalacenter/scalafix/issues/894#issue-363820370), other parts of Scalafix got faster due to the improved tree traverser currently [under review](https://github.com/scalameta/scalameta/pull/1793) in Scalameta. Rules will also benefit from that.