@olafurpg 2018-10-19
We can release scalafix v0.9.1 together with scalameta v4.1.0